backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
tensorflow==2.20.0
numpy>=2.1.0
Pillow==11.0.0